Collapsed all OptionsScreenModel methods into just two points of contact: in and out.
The previous set of methods had no reason to be so fine-grained: the
screen just called them all in a straight sequence anyway.

The new coarse-grained pair of methods expresses better when and why the
screen needs to talk to the external world: once to query data to
populate its fields, and later to return the field values so they can be
written back.

On a side note: after grouping the query methods into a single one, and
casually using JMock's "oneOf" to expect it to be called, the test
failed because the new method was invoked more than once. It
revealed a bug in OkCancelDialog that caused the panel to be
created twice.